requirements.txt
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.3.0
mediapipe>=0.10.0
opencv-python>=4.8.0
onnxruntime>=1.16.0
numpy>=1.24.0
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
eventlet>=0.35.0
sounddevice>=0.4.6
soundfile>=0.12.0
pyyaml>=6.0
psutil>=5.9.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
python-dotenv>=1.0.0
